requirements.txt
requests
google.genai
python-dotenv
openai
